Seperate classes into different files and rename namespace
